+++ src/main/java/gift/dto/request/ProductCreateRequest.java
@@ -1,4 +1,4 @@
-package gift.request;
+package gift.dto.request;
 
 import gift.validation.KakaoNotAllowed;
 import jakarta.validation.Valid;
