+++ src/main/java/gift/dto/response/CategoryResponse.java
@@ -1,4 +1,4 @@
-package gift.response;
+package gift.dto.response;
 
 public class CategoryResponse {
