+++ src/test/java/gift/service/ProductServiceTest.java
@@ -6,9 +6,9 @@ import gift.domain.Product;
 import gift.exception.ProductNotFoundException;
 import gift.repository.CategoryRepository;
 import gift.repository.ProductRepository;
-import gift.request.OptionRequest;
-import gift.request.ProductCreateRequest;
-import gift.request.ProductUpdateRequest;
+import gift.dto.request.OptionRequest;
+import gift.dto.request.ProductCreateRequest;
+import gift.dto.request.ProductUpdateRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
